refactor(challenge): Old* 레거시 파일을 challenge/old/ 폴더로 격리
- 18개 Old*.tsx 파일을 challenge/old/로 이동 (루트 17개 + ui/layout/OldChallengeLayout.tsx)
- /old/challenge/* 라우트가 운영 중이므로 삭제 대신 격리만 수행
- 외부 라우트 5곳 import 경로 갱신 (@/domain/challenge/Old → @/domain/challenge/old/Old)
- Old* 파일 내부 상대경로 정정 (./my-challenge/, ./dashboard/, ../../OldNavBar 등)
- 향후 /old/challenge/* 라우트 제거 시 challenge/old/ 폴더 통째 삭제 가능
