+++ apps/web/src/app/(user)/old/challenge/[applicationId]/[programId]/missions/[missionId]/feedback/page.tsx
@@ -1,5 +1,5 @@
 'use client';
 
-import OldMissionFeedback from '@/domain/challenge/OldMissionFeedback';
+import OldMissionFeedback from '@/domain/challenge/old/OldMissionFeedback';
 
 export default OldMissionFeedback;
