+++ apps/web/src/app/(user)/old/challenge/[applicationId]/[programId]/page.tsx
@@ -1,5 +1,5 @@
 'use client';
 
-import OldChallengeDashboard from '@/domain/challenge/OldChallengeDashboard';
+import OldChallengeDashboard from '@/domain/challenge/old/OldChallengeDashboard';
 
 export default OldChallengeDashboard;
