+++ apps/web/src/app/(user)/old/challenge/[applicationId]/[programId]/user/info/page.tsx
@@ -1,5 +1,5 @@
 'use client';
 
-import OldChallengeUserInfo from '@/domain/challenge/OldChallengeUserInfo';
+import OldChallengeUserInfo from '@/domain/challenge/old/OldChallengeUserInfo';
 
 export default OldChallengeUserInfo;
