+++ apps/web/src/domain/challenge/old/OldGuideSection.tsx
