+++ apps/web/src/domain/challenge/old/OldMissionIcon.tsx
