+++ apps/web/src/domain/challenge/old/OldMissionNotStartedIcon.tsx
