+++ apps/web/src/domain/challenge/old/OldMissionTodayIcon.tsx
