+++ apps/web/src/domain/challenge/old/OldMissionTooltipQuestion.tsx
