+++ apps/web/src/domain/challenge/old/OldNavBar.tsx
