+++ apps/web/src/domain/challenge/old/OldNoticeSection.tsx
